Cart functionality (#11)
* setting up front end cart

* cart is getting there

+++ server/db/models/friend.js
@@ -23,7 +23,7 @@ module.exports = db.define('friend', {
 		allowNull: false
 	},
 	price: {
-		type: Sequelize.DECIMAL,
+		type: Sequelize.INTEGER,
 		allowNull: false
 	},
 	imageUrl: {
